third_party/blink/web_tests/http/tests/inspector-protocol/emulation/emulation-sensor-overrides-persist-across-navigations-expected.txt

Tests that sensor overrides persist across navigations
PASS: Created gyroscope virtual sensor
Navigating to a different URL
Creating Gyroscope object in script
Verifying sensor override information
PASS: 'gyroscope' is being overridden
